% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_transcript.R
\name{extract_transcript}
\alias{extract_transcript}
\title{helper function that extracts the transcripts from a episode url including the speaker, paragraph no and the last chnage and episode lenth.}
\usage{
extract_transcript(.episode_url)
}
\arguments{
\item{.episode}{full url of an episode}
}
\value{
tibble with transcripts of episode provided
}
\description{
helper function that extracts the transcripts from a episode url including the speaker, paragraph no and the last chnage and episode lenth.
}
\examples{
extract_transcript("https://www.ndr.de/nachrichten/info/76-Coronavirus-Update-AstraZeneca-Impfstoff-besser-als-sein-Ruf,podcastcoronavirus288.html")
}
